# docs/swagger.yaml
definitions:
  dto.LoginDTO:
    properties:
      phone:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - phone
    type: object
  dto.VerifyLoginOTP:
    properties:
      otp:
        type: string
      phone:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - otp
    - phone
    type: object
info:
  contact: {}
  description: OTP-based auth service with users listing
  title: Dekamond Auth Challenge API
  version: "1.0"
paths:
  /api/v1/auth/request-otp:
    post:
      consumes:
      - application/json
      description: Generates and sends OTP for the given phone number
      parameters:
      - description: Login DTO
        in: body
        name: loginDTO
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: Request login OTP
      tags:
      - Auth
  /api/v1/auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies OTP, creates user if needed, and returns JWT
      parameters:
      - description: Verify Login Otp
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyLoginOTP'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      summary: Verify login OTP
      tags:
      - Auth
  /api/v1/users:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Search by phone
        in: query
        name: search
        type: string
      - description: Created from (RFC3339)
        in: query
        name: created_from
        type: string
      - description: Created to (RFC3339)
        in: query
        name: created_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
  /api/v1/users/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get user by id
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"